moving test to reflect code change which is already merged
